+++ test/Pool.spec.ts
@@ -91,7 +91,7 @@ describe("Pool", function () {
       poolImpl.address,
       poolImpl.interface.encodeFunctionData("initialize", [
         ethers.utils.defaultAbiCoder.encode(
-          ["address", "address", "uint256", "uint64[]", "uint64[]", "tuple(uint64, uint64)"],
+          ["address", "address", "uint32", "uint64[]", "uint64[]", "tuple(uint64, uint64)"],
           [
             nft1.address,
             tok1.address,
